Implement express tests using jest
Modify server/index.js by making a new file app.js and separating the app from the server. The reason for this is to prevent the error 'EADDRINUSE', which happens when a server is already running on the specified port number.

Write two tests to test express.js. The first test is testing for the correct status code when the get request handler is invoked with a valid restaurant id. The second test is testing to see if the get request handler retrieves the correct information given the restaurant id.
