+++ package.json
@@ -19,7 +19,8 @@
     "jquery": "^3.3.1",
     "mysql": "^2.16.0",
     "react": "^16.4.1",
-    "react-dom": "^16.4.1"
+    "react-dom": "^16.4.1",
+    "supertest": "^3.1.0"
   },
   "devDependencies": {
     "babel-core": "^6.26.3",
